requirements.txt
# This file collects all required third-party applications that are needed
# to run this project. Later you can install all these apps in a row
# using pip. Example::
#
#     pip install -r requirements.pip
Flask==0.10.1
Flask-Script==2.0.5
Flask-Admin==1.2.0
Flask-Security==1.7.4
Flask-Environments==0.1
Flask-SQLAlchemy==2.0
flask-marshmallow==0.6.2
marshmallow-sqlalchemy==0.4.1

freezegun==0.3.4
oath==1.4.1
psycopg2==2.6.1
hashids==1.1.0
validate-email==1.3
pytest==2.7.0